fix: need to group by country and state
France and other countries have overseas department that should be
seperate items in the list. They are registered as states so we need
to group them correctly and use the ISO 3166 as the key so we are able
to sort the "regular" states for the country under the correct
country.

Still one problem left as the country name is still for example
France, that need to be changed for the correct oversea department.